+++ app/src/main/res/layout/fragment_create_adopt_post.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".akbar.CreateAdoptPostFragment"> <androidx.constraintlayout.widget.ConstraintLayout
+    tools:context=".akbar.fitur_createOpenAdopt.CreateAdoptPostFragment"> <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:padding="16dp">
