+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".akbar.HomeFragment">
+    tools:context=".akbar.fitur_HomePage.HomeFragment">
 
     <de.hdodenhof.circleimageview.CircleImageView
         android:id="@+id/profile_image_home_top"
